Python program used to manage ICECAST2 dump files
This program will watch for the creation of an Icecast2 dump file as in
the case of a live broadcast. This program monitors the file for
broadcast completion and then renames and moves the file to an archive
directory using the show title queried from the Icecast server and a
time/date stamp.